Seperated view and controller a bit more
I seperated the view and the controller stronger and put the open process in a background thread. I want to try to implement the Observer Pattern tommorow and with that the correct implementation for save and open. But for this we maybe need to reconsider the structure of our programm at the moment, since we've got mostly static methods which are incompatible with the Observer class. Please give input on what you think about that. Additionally: Please use the given clean up stuff from JSBML otherwise the code comparison is quite difficult to maintain.