It is now possible to open GZip packed files #37
In this commit Issue #37 was solved, this was done by creating a new method which gets the file extension, and with that a seperation between .gz and .xml files is possible.
Howeve im not sure what the "command-line options for direct opening of files" meant, so this could be taken care of later.

Also a new test file was added, the Recon 3D model.